pyproject.toml
[tool.poetry]
name = "test-project"
version = "0.1.0"
description = ""
authors = ["Pavel Mazikin <pmazikin@gmail.com>"]
readme = "README.md"
package-mode = false
[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
pydantic-settings = "^2.6.1"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
uvicorn = "^0.32.1"
asyncpg = "^0.30.0"
pytest = "^8.3.3"
httpx = "^0.27.2"
pytest-dotenv = "^0.5.2"
pytest-asyncio = "^0.24.0"
aiosqlite = "^0.20.0"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
